pyproject.toml
[project]
name = "mis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-storage-file>=2.1.0",
    "azure-storage-file-datalake>=12.20.0",
    "cx-oracle>=8.3.0",
    "dotenv>=0.9.9",
    "ipython>=9.2.0",
    "oracledb>=3.1.0",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyspark>=3.5.5",
    "simple-ddl-parser>=1.7.1",
    "sqlglot[rs]>=26.17.1",
    "sqlparse>=0.5.3",
]
